Conflicts:
	project/target/m2e-wtp/web-resources/META-INF/maven/org.springframework.samples.service.service/project/pom.properties

+++ project/src/main/java/spring/db/lecture/LectureDao.java
@@ -155,13 +155,14 @@ public class LectureDao {
 		jdbcTemplate.update(sql, args);
 	}
 
-	public void assess(int no, Assess assess) {
+	public void assess(int no, String nick, Assess assess) {
 		String sql = "insert into assess values(?, ?, ?, ?, ?)";
-		
 		jdbcTemplate.update(sql, no, assess.getKin_grade(), assess.getPrice_grade(), assess.getKind_grade(), assess.getDetail());
 		
-		sql = "select * from assess where no = ?";
+		sql = "update teacher set students = students + 1 where teacherno = (select teacherno from lecture_info where no = ?)";
+		jdbcTemplate.update(sql, no);
 		
+		sql = "select * from assess where no = ?";
 		List<Assess> list = jdbcTemplate.query(sql, new Object[] {no}, mapper2);
 		
 		double kin_grade = 0.0;
@@ -182,8 +183,10 @@ public class LectureDao {
 		kind_grade /= (double) size;
 		
 		sql = "update lecture_info set kin_grade = ?, price_grade = ?, kind_grade = ? where no = ?";
-		
 		jdbcTemplate.update(sql, df.format(kin_grade), df.format(price_grade), df.format(kind_grade), no);
+		
+		sql = "update mylecture set eval = '평가 완료' where no = ? and id = ?";
+		jdbcTemplate.update(sql, no, nick);
 	}
 
 	public void video(int no, String title, String filename, String originalFilename, String contentType, long size) {
@@ -199,6 +202,7 @@ public class LectureDao {
 	}
 	
 	public void end() {
+		System.out.println("end 부름");
 		String sql = "select * from lecture_info where state = '등록 가능'";
 		
 		List<LectureInfo> list = jdbcTemplate.query(sql, mapper);
@@ -220,6 +224,7 @@ public class LectureDao {
 	}
 
 	public void clean() {
+		System.out.println("clean 부름");
 		String sql = "select * from lecture_info where state = '마감'";
 		
 		List<LectureInfo> list = jdbcTemplate.query(sql, mapper);
@@ -237,6 +242,16 @@ public class LectureDao {
 				
 				jdbcTemplate.update(sql, info.getNo());
 				
+				// 강사 강의 횟수 증가
+				sql = "select count(*) from lecture_info where teacherno = ?";
+				
+				int count = jdbcTemplate.queryForObject(sql, new Object[] {info.getTeacherno()}, Integer.class);
+				
+				sql = "update teacher set count = ? where teacherno = ?";
+				
+				jdbcTemplate.update(sql, count, info.getTeacherno());
+				
+				
 				sql = "delete lecture_video where no = ?";
 				
 				jdbcTemplate.update(sql, info.getNo());
@@ -247,6 +262,7 @@ public class LectureDao {
 	}
 	
 	private void deleteFile(String filename) {
+		System.out.println("파일삭제");
 		File f = new File("/resource/file/lectureVideo");
 		     
 		String fileList[] = f.list(new FilenameFilter() {
@@ -258,10 +274,56 @@ public class LectureDao {
 		 
 		});
 		
-	    File file;
-		for (int i = 0; i < fileList.length; i++) {
-			file = new File("/resource/file/lectureVideo", fileList[i]);
-			file.delete();
+		if (fileList != null) {
+		    File file;
+			for (int i = 0; i < fileList.length; i++) {
+				file = new File("/resource/file/lectureVideo", fileList[i]);
+				boolean result = file.delete();
+				System.out.println("r : " + result);
+			}
+		}
+	}
+
+	public int videoCount(int no) {
+		String sql = "select count(*) from lecture_video where no = ?";
+
+		return jdbcTemplate.queryForObject(sql, new Object[] {no}, Integer.class);
+	}
+
+	public void addVideo(int no, String title, String filename, String realname, String contentType, long size) {
+		String sql = "insert into lecture_video values(?, ?, ?, ?, ?, ?)";
+		
+		jdbcTemplate.update(sql, no, title, filename, realname, contentType, size);
+	}
+
+	public void editVideo(String filename, String title) {
+		String sql = "update lecture_video set title = ? where filename = ?";
+		
+		jdbcTemplate.update(sql, title, filename);
+	}
+
+	public void deleteVideo(int no, String filename) {
+		String sql = "delete lecture_video where filename = ?";
+		
+		jdbcTemplate.update(sql, filename);
+		
+		String[] fname = filename.split(".");
+		System.out.println("filename : " + filename);
+		System.out.println("f : " + fname[0]);
+		deleteFile(fname[0]);
+		
+		sql = "select * from lecture_video where no = ? order by filename";
+		
+		List<LectureVideo> list = jdbcTemplate.query(sql, new Object[] {no}, mapper3);
+		
+		int count = 1;
+		for (LectureVideo l : list) {
+			sql = "update lecture_video set filename = ? where filename = ? order by filename";
+			
+			String[] ext = l.getFilename().split(".");
+			jdbcTemplate.update(sql, no + "(" + count + ")." + ext[ext.length - 1], l.getFilename());
+			
+			count++;
 		}
 	}
